requirements.txt
# Meta-Lingo Backend Dependencies
# Python 3.12
# All versions are pinned to the exact versions in meta-lingo-electron environment

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
starlette==0.35.1

# HTTP Client
httpx==0.26.0

# Data Processing
numpy==2.3.5
pandas==2.3.3

# NLP - Basic
nltk==3.8.1

# Keyword Extraction
yake==0.6.0
rake-nltk==1.0.6

# NLP - SpaCy
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5

# NLP - Constituency Parsing
benepar==0.2.0
torch-struct==0.5

# Deep Learning - PyTorch
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1

# AI/ML - Transformers
transformers==4.57.3
tokenizers==0.22.1
safetensors==0.7.0
accelerate==1.12.0
huggingface-hub==0.36.0

# Video Processing - YOLO
ultralytics==8.3.237
opencv-python==4.11.0.86

# Audio/Video Processing - FFmpeg
imageio-ffmpeg==0.6.0

# Utilities
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.0
pillow==12.0.0

# Inter-coder Reliability Analysis
# krippendorff==0.8.2  # 已移除：Krippendorff's Alpha 使用手动实现，不再依赖此库
scikit-learn==1.8.0
scipy==1.16.3

# Topic Modeling - BERTopic
bertopic==0.17.4
sentence-transformers==5.2.0
umap-learn==0.5.9.post2
hdbscan==0.8.41
numba==0.63.1
llvmlite==0.46.0
pynndescent==0.5.13

# Topic Modeling - LDA (Gensim)
gensim==4.4.0
smart_open==7.5.0

# Topic Modeling - LDA Visualization
pyLDAvis==3.4.1

# Chinese Text Processing
jieba==0.42.1

# Semantic Tagging - PyMUSAS
pymusas==0.3.0

# Semantic Tagging - PyMUSAS Neural
# Neural network based semantic tagger using PyMUSAS-Neural-Multilingual-Base-BEM model
# Install with: pip install wsd-torch-models
wsd-torch-models>=0.1.0

# Visualization
plotly==6.5.0
matplotlib==3.10.8
wordcloud==1.9.3

# CLI and Utilities
typer==0.20.0
rich==14.2.0
click==8.0.4
tqdm==4.67.1

# Other Dependencies
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
filelock==3.20.0
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
networkx==3.6.1
packaging==25.0
protobuf==6.33.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
typing_extensions==4.15.0
urllib3==2.6.2
wrapt==2.0.1

# PyInstaller for packaging
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10

# SpaCy Models (install from local whl files)
# ./models/en_core_web_lg-3.8.0-py3-none-any.whl
# ./models/zh_core_web_lg-3.8.0-py3-none-any.whl

# PyMUSAS Models (install from local whl files)
# ./en_dual_none_contextual-0.3.3-py3-none-any.whl
# ./cmn_dual_upos2usas_contextual-0.3.3-py3-none-any.whl
